UX tweaks for screen reader users (#214)
* Rephrase some content Descriptions.
It is common practice to avoid including widget type or role in the description as the fact it's a button, an image is already provided by the accessibility service. At the other hand it's fine to
describe the actions or states if there is not a different mean for expressing those. It elliminates some over-verbosity.

* Dynamically update content description for the playback button

* Remove redundant content descriptions from lists.
Some accessibility services prefer content description over the text thus giving static content description for an item with dynamically updated text is not a good idea.
Additionally remove content description for the station icon. The fact it's an image is reported by the accessibility service and saying it has something to do with station in the list of stations is a
bit more verbose.

* Remove one more occurence of content description for the station icon and tweak content description for playing indicator in a list of stations