Better presentation of expanded player sheet to accessibility services (#215)
* Better presentation of expanded player sheet to accessibility services
This is very tricky to explain.
Player sheet layout is focusable and clickable. This is sufficient for
accessibility services to present it as seperate control.
When it's collapsed there are no focusable controls inside and the only
visible text it contains aggregated from all its descendants is the text
of station metadata. So it's a good idea not to override that text by
specyfying contentDescription.
When the player sheet is expanded we have a problem. It remains
clickable and focusable and has other actionable controls as its
descendants e.g. sleep timer button, copy buttons, station metadata and
station link views.
The fact player sheet is focusable and clickable determines it to be
presented to screen reader users. Accessibility services thus have to
calculate its contentDescription by aggregating texts from all its non
focusable and non actionable descendants. This automatic
contentDescription calculation is something we would like to avoid. Thus
we have to set our own contentDescription only when the sheet is
expanded.
Adding our meaningfull contentDescription also allows explaining that
clicking it will minimize the player.
If we would not use our own contentDescription accessibility services
would generate one for us by concatenating texts and contentDescriptions
from all the non focusable and non actionable descendants including
simple views holding info such as mime type, channel cound and sample
rate. Metadata and link are not included in this calculation because
these are actionable them-selves.
So if presentation of those non actionable views as a part of the
contentDescription calculation is skipped we have to make them
actionable or focusable in order to allow reading them individually by
the accessibility services.
This is the best compromise I was able to come up with. And I like the
presentation by using Talkback now.
Hopefully making those views focusable won't ruin aestetics.

Changes:
* Elliminated setting contentDescription for textual widgets
* Set labelFor property for the headings labelling information views such as station metadata, stream URL, mime type, channel count and sample rate
* Make some views focusable
* Added contentDescription for the expanded player sheet

* Fix: Care to always remove the contentDescription from the player sheet when minimizing it.